+++ R/interval_score.R
@@ -35,7 +35,11 @@
 #' @param weigh if TRUE, weigh the score by alpha / 4, so it can be averaged
 #' into an interval score that, in the limit, corresponds to CRPS. Default:
 #' FALSE.
-#' @return vector with the scoring values
+#' @param separate_results if TRUE (default is FALSE), then the separate parts
+#' of the interval score (sharpness, penalties for over- and under-prediction
+#' get returned as separate elements of a list)
+#' @return vector with the scoring values, or a list with separate entries if
+#' \code{separate_results} is TRUE.
 #' @examples
 #' true_values <- rnorm(30, mean = 1:30)
 #' interval_range = 90
@@ -56,27 +60,43 @@
 #' Johannes Bracher, Evan L. Ray, Tilmann Gneiting and Nicholas G. Reich,
 #' <arXiv:2005.12881v1>
 #'
+#' Bracher J, Ray E, Gneiting T, Reich, N (2020) Evaluating epidemic forecasts
+#' in an interval format. \url{https://arxiv.org/abs/2005.12881}
+#'
 
 
 interval_score <- function(true_values,
                            lower,
                            upper,
                            interval_range = NULL,
-                           weigh = FALSE) {
+                           weigh = TRUE,
+                           separate_results = FALSE) {
 
   if(is.null(interval_range)) {
     stop("must provide a range for your prediction interval")
   }
 
-  alpha = (100 - interval_range) / 100
+  alpha <- (100 - interval_range) / 100
 
-  score = (upper - lower) +
-    2/alpha * (lower - true_values) * (true_values < lower) +
-    2/alpha * (true_values - upper) * (true_values > upper)
+  sharpness <- (upper - lower)
+  overprediction <- 2/alpha * (lower - true_values) * (true_values < lower)
+  underprediction <- 2/alpha * (true_values - upper) * (true_values > upper)
 
-  if (weigh) score <- score * alpha / 2
 
-  return(score)
+
+  if (weigh) {
+    sharpness <- sharpness * alpha / 2
+    underprediction <- underprediction * alpha / 2
+    overprediction <- overprediction * alpha / 2
+  }
+
+  score <- sharpness + underprediction + overprediction
+
+  if (separate_results) {
+    return(list(sharpness = sharpness,
+                underprediction = underprediction,
+                overprediction = overprediction))
+  } else {
+    return(score)
+  }
 }
-
-