+++ R/pit.R
@@ -176,8 +176,8 @@ pit <- function(true_values,
     out <- list(p_value = p_value,
                 sd = NA)
 
-    if (plot == TRUE) {
-      hist_PIT <- hist_PIT(P_x, num_bins = num_bins)
+    if (plot) {
+      hist_PIT <- hist_PIT(P_x, num_bins = num_bins, caption = p_value)
       out$hist_PIT = hist_PIT
     }
 
@@ -216,8 +216,9 @@ pit <- function(true_values,
     }
 
 
-    if (plot == TRUE) {
-      hist_PIT <- hist_PIT(rowMeans(u), num_bins = num_bins)
+    if (plot) {
+      hist_PIT <- hist_PIT(rowMeans(u), num_bins = num_bins,
+                           caption = mean(p_values))
       out$hist_PIT = hist_PIT
     }
   }
@@ -238,22 +239,68 @@ pit <- function(true_values,
 #'
 #' @param PIT_samples A vector with the PIT values of size n
 #' @param num_bins the number of bins in the PIT histogram.
+#' @param caption provide a caption that gets passed to the plot
 #' If not given, the square root of n will be used
 #' @return vector with the scoring values
-#' @importFrom graphics hist
+#' @importFrom ggplot2 ggplot aes xlab ylab geom_histogram stat
 
 
-hist_PIT <- function(PIT_samples, num_bins = NULL) {
+hist_PIT <- function(PIT_samples,
+                     num_bins = NULL,
+                     caption = NULL) {
 
   if (is.null(num_bins)) {
     n <- length(PIT_samples)
     num_bins = round(sqrt(n))
   }
 
-  PIT <- PIT_samples
-  hist_PIT <- hist(PIT_samples, breaks = num_bins)
+  hist_PIT <- ggplot2::ggplot(data = data.frame(x = PIT_samples),
+                  ggplot2::aes(x = x)) +
+    ggplot2::geom_histogram(ggplot2::aes(y = stat(count) / sum(count)),
+                            breaks = seq(0, 1, length.out = num_bins + 1),
+                            colour = "grey") +
+    ggplot2::xlab("PIT") +
+    ggplot2::ylab("Frequency") +
+    ggplot2::labs(caption = paste0("p-value of Andersen-Darling test for uniformity: ",
+                                   round(caption, 3)))
 
   return(hist_PIT)
 }
 
+
+
+#' @title PIT Histogram Quantile
+#'
+#' @description
+#' Make a simple histogram of the probability integral transformed values to
+#' visually check whether a uniform distribution seems likely.
+#'
+#' @param PIT_samples A vector with the PIT values of size n
+#' @param num_bins the number of bins in the PIT histogram.
+#' @param caption provide a caption that gets passed to the plot
+#' If not given, the square root of n will be used
+#' @return vector with the scoring values
+#' @importFrom ggplot2 ggplot aes xlab ylab geom_histogram stat
+
+
+hist_PIT_quantile <- function(PIT_samples,
+                     num_bins = NULL,
+                     caption = NULL) {
+
+  if (is.null(num_bins)) {
+    n <- length(PIT_samples)
+    num_bins = round(sqrt(n))
+  }
+
+  hist_PIT <- ggplot2::ggplot(data = data.frame(x = PIT_samples),
+                              ggplot2::aes(x = x)) +
+    ggplot2::geom_histogram(ggplot2::aes(y = stat(count) / sum(count)),
+                            breaks = seq(0, 1, length.out = num_bins + 1),
+                            colour = "grey") +
+    ggplot2::xlab("PIT") +
+    ggplot2::ylab("Frequency") +
+    ggplot2::labs()
+
+  return(hist_PIT)
+}