+++ NEWS.md
@@ -1,3 +1,44 @@
+## scoringutils 0.1.4
+
+### Feature updates
+- we added basic plotting functionality to visualise scores. You can now
+easily obtain diagnostic plots based on scores as produced by `eval_forecasts`.
+- `correlation_plot` shows correlation between metrics
+- `range_plot` shows contribution of different prediction intervals to some 
+chosen metric
+- `score_heatmap` visualises scores as heatmap
+- `score_table` shows a coloured summary table of scores
+
+### package updates
+- renamed "calibration" to "coverage"
+- renamed "true_values" to "true_value" in data.frames
+- renamed "predictions" to "prediction" in data.frames
+- renamed "is_overprediction" to "overprediction"
+- renamed "is_underprediction" to "underprediction"
+
+## scoringutils 0.1.3
+
+### (Potentially) Breaking changes
+- the by argument in `eval_forecasts` now has a slightly changed meaning. It 
+now denotes the lowest possible grouping unit, i.e. the unit of one observation
+and needs to be specified explicitly. The default is now `NULL`. The reason for
+this change is that most metrics need scoring on the observation level and this 
+the most consistent implementation of this principle. The pit function receives
+its grouping now from `summarise_by`. In a similar spirit, `summarise_by` has to
+be specificed explicitly and e.g. doesn't assume anymore that you want 'range'
+to be included. 
+- for the interval score, `weigh = TRUE` is now the default option.
+- (potentially planned) rename true_values to true_value and predictions to prediction. 
+
+### Feature updates
+- updated quantile evaluation metrics in `eval_forecasts`. Bias as well as 
+calibration now take all quantiles into account
+- Included option to summarise scores according to a `summarise_by` argument in 
+`eval_forecasts` The summary can return the mean, the standard deviation as well
+as an arbitrary set of quantiles. 
+- `eval_forecasts` can now return pit histograms. 
+- switched to ggplot2 for plotting
+
 ## scoringutils 0.1.2
 
 ### (Potentially) Breakting changes